[MISC] deprecate views::take
• Moved file `range/views/take.hpp` → `io/detail/take_view.hpp`
• Moved scope `views::take` → io::detail::take`
• Replaced some `seqan3::views::take` with `std::views::take`
