+++ include/seqan3/alignment/pairwise/detail/pairwise_alignment_algorithm_banded.hpp
@@ -16,8 +16,8 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/alignment/pairwise/detail/pairwise_alignment_algorithm.hpp>
+#include <seqan3/io/detail/take_view.hpp>
 #include <seqan3/range/views/drop.hpp>
-#include <seqan3/range/views/take.hpp>
 #include <seqan3/utility/views/slice.hpp>
 
 namespace seqan3::detail
