+++ include/seqan3/utility/views/slice.hpp
@@ -19,11 +19,11 @@
 #include <stdexcept>
 #include <seqan3/std/type_traits>
 
+#include <seqan3/io/detail/take_view.hpp>
 #include <seqan3/io/exception.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/range/views/drop.hpp>
-#include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 
 namespace seqan3::detail
