+++ test/performance/range/views/view_take_benchmark.cpp
@@ -14,7 +14,7 @@
 
 #include <benchmark/benchmark.h>
 
-#include <seqan3/range/views/take.hpp>
+#include <seqan3/io/detail/take_view.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/utility/views/single_pass_input.hpp>
