+++ test/unit/io/detail/take_view_test.cpp
@@ -15,9 +15,9 @@
 #include <string>
 #include <vector>
 
+#include <seqan3/io/detail/take_view.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/container/concept.hpp>
-#include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/expect_same_type.hpp>
