+++ test/unit/range/views/CMakeLists.txt
@@ -8,7 +8,6 @@ seqan3_test(drop_test.cpp)
 seqan3_test(istreambuf_test.cpp)
 seqan3_test(move_test.cpp)
 seqan3_test(type_reduce_test.cpp)
-seqan3_test(take_test.cpp)
 seqan3_test(take_until_test.cpp)
 seqan3_test(to_lower_test.cpp)
 seqan3_test(to_upper_test.cpp)
