+++ test/unit/utility/views/pairwise_combine_test.cpp
@@ -14,7 +14,6 @@
 
 #include <seqan3/core/detail/debug_stream_tuple.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/range/views/take.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/pretty_printing.hpp>
 #include <seqan3/utility/views/pairwise_combine.hpp>
